update-child

+++ Controller/expensesmasterController.js
@@ -1,102 +1,6 @@
 const ExpensesMaster = require("../Model/expensesModel");
 const ChildExpenses = require("../Model/ChildExpensesModel");
 
-// exports.upsertExpense = async (req, res) => {
-//   //#swagger.tags = ['Master-Expenses']
-//   const { id, title } = req.body;
-//   try {
-//     if (id) {
-//       const updatedExpense = await ExpensesMaster.findByIdAndUpdate(
-//         id,
-//         { title, active: true }, // Set active to true when updating
-//         { new: true, upsert: true }
-//       );
-//       res.status(201).json({
-//         statusCode: "0",
-//         data: updatedExpense,
-//         message: "MasterExpenses Updated Successfully",
-//       });
-//     } else {
-//       const newExpensesMaster = await new ExpensesMaster({ title, active: true }).save(); // Set active to true when adding new
-//       res.status(201).json({
-//         statusCode: "0",
-//         data: newExpensesMaster,
-//         message: "MasterExpenses Added Successfully",
-//       });
-//     }
-//   } catch (error) {
-//     res.status(500).json({ statusCode: "1", message: error.message });
-//   }
-// };
-
-// exports.getAll = async (req, res) => {
-//   //#swagger.tags = ['Master-Expenses']
-//   try {
-//     const expenses = await ExpensesMaster.find();
-//     res.status(200).json({
-//       statusCode: "0",
-//       message: "Expenses data retrived Successfully",
-//       data: expenses,
-//     });
-//   } catch (error) {
-//     res.status(500).json({
-//       statusCode: "1",
-//       message: "Failed to retrived expenses data",
-//     });
-//   }
-// };
-
-// exports.getById = async (req, res) => {
-//   //#swagger.tags = ['Master-Expenses']
-//   try {
-//     const expenses = await ExpensesMaster.findOne({
-//       _id: req.params.expenses_id,
-//     });
-//     if (expenses) {
-//       res.status(201).json({
-//         statusCode: "0",
-//         message: "expenses Id retrived successfully",
-//         data: expenses,
-//       });
-//     } else {
-//       res.status(200).json({
-//         message: "expenses Id not found",
-//       });
-//     }
-//   } catch (error) {
-//     res.status(500).json({
-//       message: "Failed to retrived expenses data",
-//     });
-//   }
-// };
-
-// exports.deleteById = async (req, res) => {
-//   //#swagger.tags = ['Master-Expenses']
-//   try {
-//     const expenses = await ExpensesMaster.findById(req.params.expenses_id);
-
-//     if (expenses) {
-//       await ChildExpenses.deleteMany({ expensesId: req.params.expenses_id });
-
-//       await ExpensesMaster.findByIdAndDelete(req.params.expenses_id);
-
-//       res.status(201).json({
-//         message: "Expenses data and related ChildExpenses deleted successfully",
-//       });
-//     } else {
-//       res.status(404).json({
-//         message: "No expenses data found",
-//       });
-//     }
-//   } catch (error) {
-//     res.status(500).json({
-//       message: "Failed to delete expenses",
-//       error: error.message,
-//     });
-//   }
-// };
-
-
 exports.upsertExpense = async (req, res) => {
   //#swagger.tags = ['Master-Expenses']
   const { id, title } = req.body;
@@ -104,7 +8,7 @@ exports.upsertExpense = async (req, res) => {
     if (id) {
       const updatedExpense = await ExpensesMaster.findByIdAndUpdate(
         id,
-        { title, active: true }, // Set active to true when updating
+        { title, active: true },
         { new: true, upsert: true }
       );
       res.status(201).json({
@@ -113,7 +17,10 @@ exports.upsertExpense = async (req, res) => {
         message: "MasterExpenses Updated Successfully",
       });
     } else {
-      const newExpensesMaster = await new ExpensesMaster({ title, active: true }).save(); // Set active to true when adding new
+      const newExpensesMaster = await new ExpensesMaster({
+        title,
+        active: true,
+      }).save(); 
       res.status(201).json({
         statusCode: "0",
         data: newExpensesMaster,
@@ -128,16 +35,16 @@ exports.upsertExpense = async (req, res) => {
 exports.getAll = async (req, res) => {
   //#swagger.tags = ['Master-Expenses']
   try {
-    const expenses = await ExpensesMaster.find({ active: true }); // Only retrieve active expenses
+    const expenses = await ExpensesMaster.find();
     res.status(200).json({
       statusCode: "0",
-      message: "Expenses data retrieved successfully",
+      message: "Expenses data retrived Successfully",
       data: expenses,
     });
   } catch (error) {
     res.status(500).json({
       statusCode: "1",
-      message: "Failed to retrieve expenses data",
+      message: "Failed to retrived expenses data",
     });
   }
 };
@@ -147,7 +54,7 @@ exports.getById = async (req, res) => {
   try {
     const expenses = await ExpensesMaster.findOne({
       _id: req.params.expenses_id,
-      active: true, // Check if the expense is active
+      active: true,
     });
     if (expenses) {
       res.status(200).json({
@@ -173,16 +80,19 @@ exports.deleteById = async (req, res) => {
     const expenses = await ExpensesMaster.findById(req.params.expenses_id);
 
     if (expenses) {
-      // Instead of deleting, mark the expense as inactive
-      expenses.active = false; // Set active to false
-      await expenses.save(); // Save the changes
+      expenses.active = !expenses.active;
+      await expenses.save();
 
-      // Optionally, delete related child expenses if required
-      await ChildExpenses.deleteMany({ expensesId: req.params.expenses_id });
-
-      res.status(200).json({
-        message: "Expenses data marked as deleted successfully",
-      });
+      if (!expenses.active) {
+        await ChildExpenses.deleteMany({ expensesId: req.params.expenses_id });
+        res.status(201).json({
+          message: "Expenses data marked as deleted successfully",
+        });
+      } else {
+        res.status(201).json({
+          message: "Expenses data reactivated successfully",
+        });
+      }
     } else {
       res.status(404).json({
         message: "No expenses data found",
@@ -190,7 +100,7 @@ exports.deleteById = async (req, res) => {
     }
   } catch (error) {
     res.status(500).json({
-      message: "Failed to delete expenses",
+      message: "Failed to update expenses status",
       error: error.message,
     });
   }
